fix: build process
